backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
scikit-learn
python-multipart
numpy
streamlit
psycopg2-binary
sqlalchemy
python-dotenv
alembic
